# perf: improve performance of nested cache store