--- a/testutil/context.go
+++ b/testutil/context.go
@@ -24,7 +24,7 @@ func DefaultContext(key storetypes.StoreKey, tkey storetypes.StoreKey) sdk.Conte
 	if err != nil {
 		panic(err)
 	}
-	ctx := sdk.NewContext(cms, tmproto.Header{}, false, log.NewNopLogger())
+	ctx := sdk.NewContext(cms.CacheMultiStore(), tmproto.Header{}, false, log.NewNopLogger())
 
 	return ctx
 }
@@ -43,7 +43,7 @@ func DefaultContextWithDB(t *testing.T, key storetypes.StoreKey, tkey storetypes
 	err := cms.LoadLatestVersion()
 	assert.NoError(t, err)
 
-	ctx := sdk.NewContext(cms, tmproto.Header{}, false, log.NewNopLogger())
+	ctx := sdk.NewContext(cms.CacheMultiStore(), tmproto.Header{}, false, log.NewNopLogger())
 
 	return TestContext{ctx, db, cms}
 }