<!--
The default pull request template is for types feat, fix, or refactor.
For other templates, add one of the following parameters to the url:
- template=docs.md
- template=other.md
-->

## Description

Closes: #14377

Solution:
- Unify cachekv's cache content with single tidwall/btree.
- Use the copy-on-write supported of tidwall/btree to implement cheap `Clone`/`Restore` methods in cache store.
- Add `RunAtomic` method in Context to provide more efficient store branching out,
  it don't notifies the tracing module because it don't have the `Write` process as before.
- API breaking: Context always hold a `CacheMultiStore` instead of `MultiStore`.
- Refactor runTx and other module logics to use new `RunAtomic` method instead of `CacheContext`.

## Benchmarks

Compare `go test -run=^$ -bench=. -benchmem ./store/cachekv/... -count=5` between `1bbf3b5` and [benchmark ref](https://github.com/yihuang/cosmos-sdk/tree/benchmark_ref)(backport benchmark changes to main branch):
```
name                                 old time/op    new time/op    delta
DeepCacheStack1-12                     5.67µs ± 1%    4.53µs ± 1%   -20.17%  (p=0.008 n=5+5)
DeepCacheStack3-12                     8.00µs ± 1%    4.51µs ± 1%   -43.60%  (p=0.016 n=5+4)
DeepCacheStack10-12                    29.1µs ± 1%     4.5µs ± 1%   -84.46%  (p=0.008 n=5+5)
DeepCacheStack13-12                    48.2µs ± 1%     4.5µs ± 2%   -90.59%  (p=0.008 n=5+5)
BlankParentIteratorNextKeySize32-12    1.04µs ± 3%    0.05µs ± 2%   -94.90%  (p=0.008 n=5+5)
BlankParentAppendKeySize32-12           729ns ±13%     380ns ± 1%   -47.91%  (p=0.008 n=5+5)
SetKeySize32-12                        1.42µs ±50%    1.46µs ± 5%      ~     (p=1.000 n=5+5)
IteratorOnParentWith1MDeletes-12        1.16s ± 4%     0.00s ± 1%  -100.00%  (p=0.008 n=5+5)
CacheKVStoreGetNoKeyFound-12            708ns ± 2%     678ns ± 1%    -4.26%  (p=0.008 n=5+5)
CacheKVStoreGetKeyFound-12              176ns ±11%     336ns ± 3%   +90.47%  (p=0.008 n=5+5)
CacheKVStoreSetAndCommit-12            9.69µs ± 0%    6.12µs ± 2%   -36.85%  (p=0.016 n=4+5)

name                                 old alloc/op   new alloc/op   delta
DeepCacheStack1-12                     2.10kB ± 0%    1.74kB ± 0%   -17.18%  (p=0.008 n=5+5)
DeepCacheStack3-12                     2.83kB ± 0%    1.74kB ± 0%   -38.70%  (p=0.008 n=5+5)
DeepCacheStack10-12                    5.41kB ± 0%    1.74kB ± 0%   -67.90%  (p=0.008 n=5+5)
DeepCacheStack13-12                    6.51kB ± 0%    1.74kB ± 0%   -73.34%  (p=0.008 n=5+5)
BlankParentIteratorNextKeySize32-12      205B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
BlankParentAppendKeySize32-12            238B ±16%      112B ± 0%   -52.94%  (p=0.008 n=5+5)
SetKeySize32-12                          327B ±19%      164B ± 0%   -49.76%  (p=0.008 n=5+5)
IteratorOnParentWith1MDeletes-12        253MB ± 0%       0MB           ~     (p=0.079 n=4+5)
CacheKVStoreGetNoKeyFound-12             283B ± 1%      163B ± 0%   -42.32%  (p=0.008 n=5+5)
CacheKVStoreGetKeyFound-12              3.00B ± 0%     3.00B ± 0%      ~     (all equal)
CacheKVStoreSetAndCommit-12            4.83kB ± 0%    3.02kB ± 0%   -37.54%  (p=0.008 n=5+5)

name                                 old allocs/op  new allocs/op  delta
DeepCacheStack1-12                       19.0 ± 0%      12.0 ± 0%   -36.84%  (p=0.008 n=5+5)
DeepCacheStack3-12                       31.0 ± 0%      12.0 ± 0%   -61.29%  (p=0.008 n=5+5)
DeepCacheStack10-12                      73.0 ± 0%      12.0 ± 0%   -83.56%  (p=0.008 n=5+5)
DeepCacheStack13-12                      91.0 ± 0%      12.0 ± 0%   -86.81%  (p=0.008 n=5+5)
BlankParentIteratorNextKeySize32-12      2.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
BlankParentAppendKeySize32-12            1.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
SetKeySize32-12                          1.40 ±43%      0.00       -100.00%  (p=0.008 n=5+5)
IteratorOnParentWith1MDeletes-12        3.07M ± 0%     0.00M           ~     (p=0.079 n=4+5)
CacheKVStoreGetNoKeyFound-12             4.00 ± 0%      2.00 ± 0%   -50.00%  (p=0.008 n=5+5)
CacheKVStoreGetKeyFound-12               1.00 ± 0%      1.00 ± 0%      ~     (all equal)
CacheKVStoreSetAndCommit-12              97.0 ± 0%      45.0 ± 0%   -53.61%  (p=0.008 n=5+5)
```

Benchmark result looks pretty good, it's a huge win on nested cases and iterator cases, it's even much faster in the `SetAndCommit` case which combines several `Set` operations and `cache.Write` together. `CacheKVStoreGetKeyFound` is the major slowdown case.

<!-- Add a description of the changes that this PR introduces and the files that
are the most critical to review. -->

---

### Author Checklist

*All items are required. Please add a note to the item if the item is not applicable and
please add links to any relevant follow up issues.*

I have...

- [ ] included the correct [type prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json) in the PR title
- [ ] added `!` to the type prefix if API or client breaking change
- [ ] targeted the correct branch (see [PR Targeting](https://github.com/cosmos/cosmos-sdk/blob/main/CONTRIBUTING.md#pr-targeting))
- [ ] provided a link to the relevant issue or specification
- [ ] followed the guidelines for [building modules](https://github.com/cosmos/cosmos-sdk/blob/main/docs/docs/building-modules)
- [ ] included the necessary unit and integration [tests](https://github.com/cosmos/cosmos-sdk/blob/main/CONTRIBUTING.md#testing)
- [ ] added a changelog entry to `CHANGELOG.md`
- [ ] included comments for [documenting Go code](https://blog.golang.org/godoc)
- [ ] updated the relevant documentation or specification
- [ ] reviewed "Files changed" and left comments if necessary
- [ ] confirmed all CI checks have passed

### Reviewers Checklist

*All items are required. Please add a note if the item is not applicable and please add
your handle next to the items reviewed if you only reviewed selected items.*

I have...

- [ ] confirmed the correct [type prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json) in the PR title
- [ ] confirmed `!` in the type prefix if API or client breaking change
- [ ] confirmed all author checklist items have been addressed 
- [ ] reviewed state machine logic
- [ ] reviewed API design and naming
- [ ] reviewed documentation is accurate
- [ ] reviewed tests and test coverage
- [ ] manually tested (if applicable)