@yihuang 2023-01-20
> LGTM! @yihuang can we get this over the finish line? I would also like another look and approval from other folks. Maybe @testinginprod?

the only unsatisfactory part if the `Restore` api can't specify self type in interface, instead it receive a `types.CacheKVStore` and cast to self, but I don't know how to improve that.
other than that, I think it's done ;D