@yihuang 2023-01-20
> LGTM! @yihuang can we get this over the finish line? I would also like another look and approval from other folks. Maybe @testinginprod?

it also has impact on tracing support, since there's no `Write` process, which previously trigger something in tracing, I'm not familiar with that feature.