@yihuang 2023-01-20
@08d2 I removed the `CloneMultiStore()` method, and all the changes related to it, only keep the `RunAtomic`, appreciate the review suggestions.