@tac0turtle 2023-01-27
Been thinking about this pr. i forgot who, but someone mentioned how its api breaking in a decently sized manner and if we want to push this breakage onto users now when we may push another store breaking change on them in 2-4 months. Would like to hear from others if we think this is fine or we should aim for a single api break change instead of two with in a 2-4 month time period