@yihuang 2023-01-27
> we may push another store breaking change on them in 2-4 months

do you mean a bigger refactoring of cache store?