@alexanderbez 2023-01-28
> > we may push another store breaking change on them in 2-4 months
> 
> do you mean a bigger refactoring of cache store?

Yeah might be best to hold off until we have a more clear picture