@tac0turtle 2023-02-19
Hey, sorry for the long delay. Lets close this pr for now, and then talk about it in the next storage working group call. With the adr being open for storage, id hate to break the api on users twice instead of once. Sorry again for the long delay.